refactor: models → model 로 네미밍 변경
