+++ src/feature/todo/index.ts
@@ -1,5 +1,5 @@
 export { TodayMissionBoard } from './todayMissionBoard/component';
-export { PlanSelect } from '../../models/todo/planSelector/component';
-export { TodoListProvider, useTodoBoardState, useTodoBoardActions } from '../../models/todo/todoList/context';
+export { PlanSelect } from '../../model/todo/planSelector/component';
+export { TodoListProvider, useTodoBoardState, useTodoBoardActions } from '../../model/todo/todoList/context';
 
-export { PlanSelectorProvider, usePlanSelector } from '../../models/todo/planSelector/hooks';
+export { PlanSelectorProvider, usePlanSelector } from '../../model/todo/planSelector/hooks';
