+++ src/model/goal/context.ts
